src/requirements/base.txt
django==4.2
djangorestframework==3.14.0
django-environ==0.10.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
